Fix small issues with respect to alignment of the subtotal, printing the whiteline and sorting of tasks
Some projects do not have tasks, therefor the task column will not be printed, so there is a column less.
We still like to have the subtotalnumber printed below the duration.
This is now done in the correct column.

For the whiteline we better print a space in certain cells, instead of white text.

When all tasks are WP tasks, we need to order on the WP number instead of the task string.
We want WP10 to be printed after WP9, not WP1.