Extend filter on account analytic lines.
This is quite a generic table, and we find more than timesheet related records.

In this situation we deal with expense related records, and we do not want them to appear
on the timesheet report (printed from the timesheet wizard 'Print timesheets').

Instead of updating the query to find the timesheet related records that we need for the report, we choose otherwise:
We look for the records that we do not want to appear on the report.
There are 6 attributes that we can use to uniquely identify the expense related acc.analytic lines, given the current dataset.
Using one attribute would already be enough, but we rather use a combination as it is unclear if future data
conforms to the patterns we currently see (and expect) in the table.

Using a combination makes the (overall) filter more strict with respect to non timesheet related records.
We do not want False Negatives for the filter on timesheet related records, we rather have False Positives.
That might mean that in the future, lines might show up on the report, that do not belong there, if so, we could
update the filter (again).